+++ host-run-on-evercran.sh
@@ -1,15 +1,13 @@
-#!/bin/bash
-
 # TODO: make args capture whole args...
 # TODO: R 0.49 segfaults
 # TODO: 
 #       - no .Last.sys in 3.0.0 or 2.15.3
 #       - no last.warning in 4.2.0/1... 
-#       - no tcltk in 1.5.0-2.0.0
 #       
 
 source host-functions.sh
 
+echo "Have you started Xquartz and run 'xhost +' in a separate terminal?"
 # cleanup
 rm docker-data/*
 rm -rf opt-R-volumes/*